chore: relocate controller package to sink directory
