+++ cmd/front/main.go
@@ -5,9 +5,9 @@ package main
 
 import (
 	"github.com/sonr-io/motr/config"
-	"github.com/sonr-io/motr/controller"
 	"github.com/sonr-io/motr/middleware"
 	"github.com/sonr-io/motr/routes"
+	"github.com/sonr-io/motr/sink/controller"
 )
 
 func main() {
