+++ sink/controller/asset/entity.go
