+++ sink/controller/prices/entity.go
