+++ sink/controller/profile/inputs.templ
