+++ sink/controller/profile/inputs_templ.go
